requirement.txt
alembic==1.4.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Injector==0.12.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
greenlet==0.4.15
idna==2.8
injector==0.18.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3
requests==2.22.0
singleton-decorator==1.0.0
six==1.14.0
SQLAlchemy==1.3.13
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
